feat(interface): implement “New client” modal window
Added a modal window with a form.
When clicking on the “Add a client” button, a POST-request sent to the server.
When you click on the “Add contact” button, up to 10 inputs get with the ability to change their type and delete them.
If an error message received from the server, it displays above the “Send” button.
Inputmask library has added.
The modal window can close by clicking on the cross button, cancel button, or background button.

closes issue #2