feat(interface): implement “Change client” modal window
Added a modal window with a form.
The window opens with fields filled with the appropriate data.
When clicking on the “Save” button, a PATH-request sent to the server.
When you click on the “Add Contact” button, up to 10 inputs appear with the ability to change their type and delete them.
If an error message gets from the server, it displays above the “Send” button.
A modal window can close by clicking on the cross button, cancel button or background button.

closes issue #3